Reorder and rename components
